move theme extensions to index, reorg
