+++ packages/ui-components/src/styles/theme/interfaces.ts
@@ -1,0 +1,8 @@
+/** Theme interfaces */
+import { TypographyOptions } from "@mui/material/styles/createTypography";
+
+/** Typography */
+export interface ExtendedTypographyOptions extends TypographyOptions {
+  paragraphSmall: React.CSSProperties;
+  paragraphLarge: React.CSSProperties;
+}
