+++ packages/ui-components/src/styles/theme/theme.ts
@@ -1,28 +1,6 @@
-import { createTheme, Theme } from "@mui/material/styles";
+import { createTheme } from "@mui/material/styles";
 import palette from "./palette";
 import typography from "./typography";
-
-declare module "@mui/material/styles" {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface DefaultTheme extends Theme {}
-
-  interface Palette {
-    severity: {
-      100: string;
-      200: string;
-      300: string;
-      400: string;
-      500: string;
-    };
-    gradient: {
-      100: string;
-      200: string;
-      300: string;
-      400: string;
-      500: string;
-    };
-  }
-}
 
 /**
  * Theme configuration variables
